Imported all Card animation frames. There were about 1500 animatoin frames, and I removed almost 90% of them which were redundant.
Card Animator is now fully rigged, with the help of a script I wrote called CardImageSwapper. This scripts makes it so I didn't have to create an animation clip and Animator state for 34 different card animations; instead, I only have 2 animations and swap out "balloon" with a different string on specific frames. The script is general purpose, so could be used again if we run into this problem in the future.