Reorganized some code files. Modified some of the sunflower sprites. Sunflower netcode is in place, as well as some of the procedural animations for the flowers.
Sunflowers can now be looped between three states: EmptyMound, Sprout, and Flower, by hopping on them (the hop animation itself may be later disabled in this zone). The flower will smoothly scale up over 10 seconds, then smoothly scale down after the flower has been picked off the stalk. The flowers will be packed pretty close together, but it's a little finicky to overlap them exactly, so I gave them a large collision radius, and added logic to sort out the closest flower out of all the flowers the player is overlapping, to be the one that is interacted with (this'll make interacting feel easier and more fun).