Added realtime asset downloading with Unity's new Addressables system! This will reduce the build footprint by a huge amount, because we can easily tell almost anything (except scripts) to be downloaded from a remote server. In addition, it could be a huge boon to iteration -- On my machine, Android builds now happen in 30 seconds, as opposed to 20 minutes, which is around how long it always took when packing in all the huge atlases.
Made a decent looking loading screen where the assets are downloaded.

The Addressables are currently hosted on my personal server (ultimatewalrus.com), so we'll need to eventually determine where to host the files. In the meantime, if you're another engineer needing to iterate past this commit, just create a new profile in Addressables that bundles in the assets.


Former-commit-id: 517063074065187c65399c1d954e36b3a09fae43