Basic partner disconnect countdown is in along with proper UI.  However this will begin as soon as you join and interrupt the intro. I'm going to have to add a lot of logic after this to deal with the special case of initially waiting for your date.
Added a debug button that you can use to stop the partner disconnect countdown, if you need to for testing. Also added the "Debug Settings" asset, and scripts which delete anything debug-related for a Release build. You can configure this obejct by just searching for "Debug Settings" in assets, and changing it between Debug and Release.