Fune tuned skin selection. Non premium players can now only select from the skins indicated in the spec. Additionally, these are the pool of skins that will be chosen from if the player first joins, and has never selected a skin. If they never open the inventory and select a skin, the app will continue to pick a random one every time they join. As soon as they pick a skin, it's saved, and will be restored when they next join.
I had to rewrite the skin selection system, as it didn't seem to be working properly. Now, there is no custom array saved to the room; rather, we'll just take stock of all the used skins in a room when we join it, and select randomly among the non-premium skins which are least-used in the room.