Did a bunch of work to solve a simple problem: how do we make sure both matched players get placed in the same room? And how do we make sure another player doesn't take the spot of the match we're waiting for?
The approach I took is outlined in TeamRoomJoin.cs. In summary, players look to see if their "friend" is in a server already. If so, they join it. If not, they either join an existing room, or if none are available, they'll create one, while reserving another slot for their match. If their match is taking a while to join, they'll occasionally ping Photon to see where their friend is, to make sure they didn't accidentally join a different server.

Needs testing.


Former-commit-id: 09ccbde8c6589b19a73981bb31645ea5b2b68e62