Loading screen is now styled with a screenshot background, so it looks like you're already in Dateland, and it just has "Connecting...XX%" shown as the download occurs.
Handled errors more gracefully in this screen. Users will be told "A problem occurred. Please try again!" if an error occurs while connecting. It immediately sends DatelandfailureToConnect("Could not download or start the game.") to native. Then, after 5 seconds, it sends playerDidLeaveGame() to native. It will await the next call to createPlayer(), at which point it will restart the loading process. I don't disconnect Agora however -- if they managed to get that to connect, we may as well let them talk, at least.

Also fixed a slightly misleading popup for when the internet is disconnected on entry to Dateland. If a disconnect occurred, it'll change from "Connecting..." to "Reconnecting..." to indicate a problem is occurring. Then it will do the usual 60-second retry period before letting the player know they've been disconnected.


Former-commit-id: c5197f9a12e38a81e1f8bfe08cb2ca444acd0c71